wannacry-detector-lite/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wannacry-detector-lite"
version = "1.0.0"
description = "Lightweight WannaCry ransomware detector using ML, PE analysis, and YARA rules"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "PTIT Security Research Lab"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "pefile>=2023.2.7",
    "yara-python>=4.2.3",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "customtkinter>=5.2.0",
    "reportlab>=4.0.0",
    "joblib>=1.3.0",
    "imbalanced-learn>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pyright>=1.1.330",
]

[project.scripts]
wannacry-detector = "main:main"
wannacry-train = "train_model:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B"]
ignore = ["N806"]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
include = ["core", "gui", "main.py", "train_model.py"]
exclude = ["tests"]
pythonVersion = "3.10"
typeCheckingMode = "basic"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --cov=core --cov-report=term-missing"
